requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
chromadb==0.5.5
sentence-transformers==3.1.1
pymupdf==1.24.10
python-dotenv==1.0.1
anthropic==0.34.2
aiofiles==24.1.0
pandas==2.2.2
